Fix AI chat review findings and cleanup legacy service
